refactor(surveyResult): move LoadSurveyResult to a protocols file
